refactor(shifts): move shifts config into lib
