--- config/jest/cssTransform.cjs
@@ -1,8 +1,0 @@
-module.exports = {
-  process() {
-    return "module.exports = {};";
-  },
-  getCacheKey() {
-    return "cssTransform";
-  },
-};
