+++ lib/utils.js
@@ -41,7 +41,7 @@ export function scrollToADay(year, month, day) {
 /**
  * Get the URL for a calendar page.
  * @param {object} param   Param Object
- * @param {import("./constants").ShiftModels}  param.shiftModel    Key of the ShiftModel used.
+ * @param {import("./shifts").ShiftModelKeys}  param.shiftModel    Key of the ShiftModel used.
  * @param {boolean}           [param.isFullYear]  Should the URL be the full year page url.
  * @param {number}            [param.year]        Year of the page.
  * @param {number}            [param.month]       Month of the page. Defaults to current month.
